Refactor: JVM 내 Topic 데이터 보관 로직 Topic 디렉토리로 분리해라 - #85
