+++ src/main/java/talkPick/external/llm/dto/LLMReqDTO.java
@@ -2,7 +2,7 @@ package talkPick.external.llm.dto;
 
 import talkPick.domain.random.dto.MemberDataDTO;
 import talkPick.domain.random.dto.RandomTopicHistoryDataDTO;
-import talkPick.domain.random.dto.TopicDataDTO;
+import talkPick.domain.topic.dto.TopicDataDTO;
 import java.util.List;
 
 public record LLMReqDTO(
