+++ src/main/java/talkPick/external/llm/port/LLMClientPort.java
@@ -3,7 +3,7 @@ package talkPick.external.llm.port;
 import talkPick.domain.random.adapter.out.dto.RandomResDTO;
 import talkPick.domain.random.dto.MemberDataDTO;
 import talkPick.domain.random.dto.RandomTopicHistoryDataDTO;
-import talkPick.domain.random.dto.TopicDataDTO;
+import talkPick.domain.topic.dto.TopicDataDTO;
 
 import java.util.List;
